Moved singleFacilityOccupancy call to API, facility onResponse to FacilitiesActivity
